Fixed issue in `DrawUtil#lengthOfPath`
Fixed issue in `DrawUtil#lengthOfPath` where path close (which connects the last point to the beginning, effectively another line in the path) did not count towards the line path count.
Now, that extra line is accounted for.

The test for this method has been updated to reflect the change.
A new test was also added to ensure this method works with multiple subpaths.