Begin writing unit tests for Text2D class
This one's a doozy... testing `Text2D` methods requires a lot more work than other testing setups so far.
`Text2D` objects depend directly on graphics received from the `Canvas` object in `FastJEngine#display`. In order for this to be available, the game engine would have to be instantiated and run, and the tests would have to be performed from inside an instance of the engine.
The original state of the engine would not properly allow for this to be possible -- calling `FastJEngine#closeGame()` outright closed the application, and `Display#close()` would end up doing the same.
As such, I modified the `FastJEngine#closeGame()` method to properly allow for reusability of the game engine in one running JVM instance, and allow writing of tests for the `Text2D` class.
Now, tests can be conducted for the `Text2D` class by initializing the engine with a logic manager containing the scene with the testing needed. The engine would be run, and while the engine runs it would be told to close.

New Additions:
- Added method `DrawUtil#randomFont()` to generate random fonts
- Add default color and show constants to `Text2D`
- Added method `Maths#randomInteger()` to generate random integers

Other Changes:
- Changed constant `Text2D#DEFAULT_FONT` to `Text2D#defaultFont`.
- Minor changes to documentation all around
- Changed `FastJEngine#exit()` in two key ways:
    - Removed the call of `System.exit(0)`
    - Reset all the data contained in the engine.
  This allows the game engine to be reused after closing a game instance.